+++ LethalHands/CHANGELOG.md
@@ -1,3 +1,9 @@
+### v22.0.5 :
+	- Added configuration for punches (range, cooldown, damage, stamina drain,chance to deal damage)
+	- Adjusted some default values : 
+		- Punch range : 1f -> 1.2f (knife is 0.75f and shovel is 1.5f for reference)
+		- Stamina drain : 10% -> 0%
+
 ### v22.0.4 :
 	- Changed control tips to reflect the correct controls (any instance of [RMB] in the tips is replaced by the game with [LMB] for some reason)
 	- The mod now changes LethalEmotesAPI's third person mode to Normal upon loading, this should fix people being in third person